survey, comment, netstedcomment 의 save, get

+++ src/main/kotlin/com/motline/mushroom/repository/VoteRepository.kt
@@ -18,7 +18,7 @@ class VoteRepository {
     }
 
     suspend fun get(id: String): Vote? {
-        return this.votes.find { it.id == id.toInt() }
+        return this.votes.find { it.id == id }
     }
 
     suspend fun getAll(): List<Vote> {
